updating slice naming conventions
